Add Preporcessing Scripts and a README to the sponges tools (#96)
* Add scipts and readme for preprocessing data for sponges

Partial readme including steps for using the preprocessing scripts
and what changes need to be done in the xml to use temperature and
salinity sponges in MOM6.

Preprocessing scripts subset and average daily data from uda, fill
the monthly data, and merge the monthly T and S averages into one
file for each year.

* Update README.md

Add more details to the readme

* Update fill_glorys_nn_monthly.sh

* Claen up scripts and add paths

Add paths variables for archive and work.

* Change file extensions

Change file extensions for cshell scripts

* Remove .sh files

* Update paths to use TMPDIR

- change get_so and get_thetao to use a TMPDIR as scratch space
- remove cnk options since they aren't needed
- combine dmget statement

* Fix Error in get_thetao

A typo in get_thetao mean so was being used instead. This has been fixed and
not a variable ${var} has been added. This means that lines 18 to 29 should
be the same in get_so and get_thetao.

* Simplify merge_so_thetao

- simplify code based on YCT suggestion
- update usage comment and job name

* Update README.md

- fix typo
-change file names from .sh to .csh

* Update README.md

typos:
-fix spelling
-add word

---------

Co-authored-by: Theresa Morrison <Theresa.Morrison@an200.princeton.rdhpcs.noaa.gov>
Co-authored-by: Theresa Morrison <Theresa.Morrison@an204.princeton.rdhpcs.noaa.gov>
Co-authored-by: Theresa Morrison <Theresa.Morrison@an205.princeton.rdhpcs.noaa.gov>
Co-authored-by: Yi-Cheng Teng - NOAA GFDL <143743249+yichengt900@users.noreply.github.com>